Resolve "Refactor following FastAPI best practice"

Closes #9

See merge request dti-idsia/spearhead/federated-learning-server!22

+++ ferdelance/database/tables.py
@@ -1,4 +1,3 @@
-from email.policy import default
 from sqlalchemy import Column, ForeignKey, String, Float, DateTime, Integer, Boolean, Date
 from sqlalchemy.sql.functions import now
 from sqlalchemy.orm import relationship
@@ -47,7 +46,7 @@ class ClientToken(Base):
 
     token_id = Column(Integer, primary_key=True, autoincrement=True, index=True)
     creation_time = Column(DateTime(timezone=True), server_default=now())
-    expiration_time = Column(Integer, nullable=True)
+    expiration_time = Column(Float, nullable=True)
     token = Column(String, nullable=False, index=True, unique=True)
     valid = Column(Boolean, default=True)
